implementation-examples/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
dnspython==2.8.0
email-validator==2.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pydantic==2.12.5
pydantic-core==2.41.5
rota-connector @ git+ssh://git@bitbucket.org/31green/rota_connector.git@673f8126112cd8e8149b53aeca5cb33f14ffb3c6
typing-extensions==4.15.0
typing-inspection==0.4.2
python-dotenv==1.0.1
